requirements.txt
django
factory_boy

ipdb
